Changed render method of objects to use sf::RenderTarget. Corrected crash in PixelPerfectTest. Minor corrections.
git-svn-id: svn://localhost@535 8062f311-0dae-4547-b526-b8ab9ac864a5

+++ GDL/GDL/RuntimeScene.cpp
@@ -426,8 +426,8 @@ bool RuntimeScene::UpdateTime()
     realElapsedTime -= pauseTime; //On enlève le temps de pause
 
     //On modifie ce temps écoulé si il est trop bas.
-    if ( game->minFPS != 0 && realElapsedTime > 1000.f/static_cast<float>(game->minFPS) )
-        realElapsedTime = 1000.f/static_cast<float>(game->minFPS); //On ralentit le jeu si les FPS sont trop bas.
+    if ( game->minFPS != 0 && realElapsedTime > 1000.0/static_cast<double>(game->minFPS) )
+        realElapsedTime = 1000.0/static_cast<double>(game->minFPS); //On ralentit le jeu si les FPS sont trop bas.
 
     elapsedTime = realElapsedTime*timeScale; //Le temps écoulé par le jeu est modifié suivant l'échelle du temps
